Move Post and ParentalControlSettings components out of their folders
